fix(utils): rename constants to utils
