+++ apps/scripts/src/utils.js
